+++ pc/src/main/java/group2/sdp/pc/comms/Commands.java
@@ -9,25 +9,25 @@ package group2.sdp.pc.comms;
  * Based on code by SDP Group 4 2013
  */
 public class Commands {
-	public final static int DO_NOTHING = 0;
-	public final static int FORWARDS = 1;
-	public final static int BACKWARDS = 2;
-	public final static int STOP = 3;
-	public final static int KICK = 4;
-	public final static int DISCONNECT = 5;
-	public final static int ROTATE = 6;
-	public final static int ROTATEMOVE = 7;
-	public final static int TRAVEL_ARC = 8;
-	public final static int ACCELERATE = 9;
-	public final static int LEFT = 10;
-	public final static int RIGHT = 11;
-	public final static int ANGLEMOVE = 12;
-	public final static int SLOWMOVE = 13;
-	public final static int DRIBBLERON = 23;
-	public final static int DRIBBLEROFF = 24;
-	public final static int TEST = 66;
-	public final static int FORCEQUIT = 55;
-	public final static int BEEP = 42;
-	public final static int ARC = 37;
-	public final static int STEER = 36;
+	public final static short DO_NOTHING = 0;
+	public final static short FORWARDS = 1;
+	public final static short BACKWARDS = 2;
+	public final static short STOP = 3;
+	public final static short KICK = 4;
+	public final static short DISCONNECT = 5;
+	public final static short ROTATE = 6;
+	public final static short ROTATEMOVE = 7;
+	public final static short TRAVEL_ARC = 8;
+	public final static short ACCELERATE = 9;
+	public final static short LEFT = 10;
+	public final static short RIGHT = 11;
+	public final static short ANGLEMOVE = 12;
+	public final static short SLOWMOVE = 13;
+	public final static short DRIBBLERON = 23;
+	public final static short DRIBBLEROFF = 24;
+	public final static short TEST = 66;
+	public final static short FORCEQUIT = 55;
+	public final static short BEEP = 42;
+	public final static short ARC = 37;
+	public final static short STEER = 36;
 }
